Update sales price data type in OrderItem model

+++ server_asp/src/models/OrderItem.cs
@@ -11,7 +11,7 @@ public partial class OrderItem
 
     public int Quantity { get; set; }
 
-    public int SalesPrice { get; set; }
+    public float SalesPrice { get; set; }
 
     public DateTime CreatedAt { get; set; }
